+++ src/localisation/phrases.ts
@@ -300,6 +300,8 @@ enum Phrase {
   ShareableLinkTitle,
   ShareableLinkText,
   ShareableLinkFailedTitle,
+  VideoActionFailed,
+  DiscardChanges,
   ShareableLinkFailedText,
   CloudUsageDescription,
   DiskUsageDescription,