+++ src/main/types.ts
@@ -280,6 +280,11 @@ type RendererVideo = Metadata & {
   uniqueId: string;
 };
 
+type VideoAction =
+  | { type: 'protect'; value: boolean }
+  | { type: 'tag'; value: string }
+  | { type: 'delete' };
+
 type RendererClip = RendererVideo & { category: VideoCategory.Clips };
 
 type SoloShuffleTimelineSegment = {
@@ -725,6 +730,7 @@ export {
   VideoQueueItem,
   Metadata,
   RendererVideo,
+  VideoAction,
   RendererClip,
   Flavour,
   SoloShuffleTimelineSegment,